% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmeds.R
\name{allmeds}
\alias{allmeds}
\title{Gets matrices of medians for each individual sample for all measured parameters for all clusters}
\usage{
allmeds(citrus.combinedFCSSet, citrus.foldClustering, citrus.foldFeatureSet)
}
\arguments{
\item{citrus.combinedFCSSet}{loaded from citrusClustering.RData file generated by Citrus run}

\item{citrus.foldClustering}{loaded from citrusClustering.RData file generated by Citrus run}

\item{citrus.foldFeatureSet}{computed from first two variables using citrus.calculateFoldFeatureSet function from citrus package}
}
\value{
Returns a list with each element corresponding to a matrix (rows as samples, columns as measured parameters) for a different cluster (for the minimum threshold specified)
}
\description{
Gets matrices of medians for each individual sample for all measured parameters for all clusters
}
\examples{
library(mineCitrus)
data("citrus.combinedFCSSet")
data("citrus.foldClustering")
data("citrus.foldFeatureSet")
meds<-allmeds(citrus.combinedFCSSet=citrus.combinedFCSSet,
              citrus.foldClustering=citrus.foldClustering,
              citrus.foldFeatureSet=citrus.foldFeatureSet)

}
