% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-citrus.combinedFCSSet.R
\docType{data}
\name{citrus.combinedFCSSet}
\alias{citrus.combinedFCSSet}
\title{Cytometry data set for example of Citrus data set from nolanlab/citrus}
\format{A large citrus.combinedFCSSet object with 5 elements:
\describe{
  \item{data}{Toy data set for cytometry}
  \item{fileChannelNames}{Names of channels for measured parameters included in toy cytmetry data set}
  \item{fileIds}{ID numbers for each file included in toy cytometry data set}
  \item{fileNames}{Names of files included in toy cytmetry data set}
  \item{fileReagentNames}{Names of measured channels in toy cytmetry data set}
  ...
}}
\source{
\url{https://github.com/nolanlab/citrus}
}
\usage{
citrus.combinedFCSSet
}
\description{
A dataset containing the a simple example of cytometry data
}
\keyword{datasets}
