% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-citrus.foldFeatureSet.R
\docType{data}
\name{citrus.foldFeatureSet}
\alias{citrus.foldFeatureSet}
\title{Correlation data for example of Citrus data set from nolanlab/citrus}
\format{A list with 8 elements:
\describe{
  \item{allFeatures}{Data set for each sample for all markers and clusters}
  \item{allLargeEnoughClusters}{Vector of clusters meeting size threshold}
  \item{foldFeatures}{Data for each fold clustering}
  \item{foldLargeEnoughClusters}{Clusters meeting size threshold for each fold clustering}
  \item{folds}{Descriptions of each data clustering}
  \item{leftoutFeatures}{Data omitted from analyses}
  \item{minimumClusterSizePercent}{Minimum size threshold to retain clusters in analysis}
  \item{nFolds}{The number of times data is clustered}
  ...
}}
\source{
\url{https://github.com/nolanlab/citrus}
}
\usage{
citrus.foldFeatureSet
}
\description{
A dataset containing the association of red and blue in clusters with different sample groups
}
\keyword{datasets}
